After reading the spec:
https://www.w3.org/TR/wai-aria-1.1/#aria-hidden

It looks like aria-hidden="false" has inconsistent support across browsers, so a better approach
is not set a defualt and explicitly set its state depending on context. this seems like a better
approach from progressive enhancement standpoint. If we're on small screen there might not be
companion text w/ an icon, so this way the icon's title should be read out, b/c we're labelling and
setting it via the title it, and text between the <title> tags. The spec also recommends an
icon w/ aria-hidden="true" be hidden w/ css: [aria-hidden="true"] { visibility: hidden; }